minor corrections after code review
