+++ pkg/telemetry/phonehome/gatherer.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"time"
 
+	"github.com/pkg/errors"
 	"github.com/stackrox/rox/pkg/concurrency"
 	"github.com/stackrox/rox/pkg/sync"
 )
@@ -107,4 +108,15 @@ func (g *gatherer) AddGatherer(f GatherFunc) {
 	g.mu.Lock()
 	defer g.mu.Unlock()
 	g.gatherFuncs = append(g.gatherFuncs, f)
+}
+
+// AddTotal sets an entry in the props map with key and number of elements
+// returned by f as the value.
+func AddTotal[T any](ctx context.Context, props map[string]any, key string, f func(context.Context) ([]*T, error)) error {
+	ps, err := f(ctx)
+	if err != nil {
+		return errors.Wrapf(err, "failed to get %s", key)
+	}
+	props["Total "+key] = len(ps)
+	return nil
 }
